Refactor : ReactDOM.render와 컴포넌트 분리
