+++ src/index.jsx
@@ -1,19 +1,5 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import IncrementCounter from './IncrementCounter';
-import CounterContainer from './CounterContainer';
-
-const App = () => {
-  const [count, setCount] = useState(0);
-  return (
-    <>
-      <div>Counter</div>
-      <IncrementCounter count={count} setCount={setCount} />
-      <CounterContainer setCount={setCount} />
-    </>
-  );
-};
-
-export default App;
+import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('app'));
